review_eval/pyproject.toml
[project]
name = "review-eval"
version = "0.1.0"
description = "Golden test suite to verify Claude Code review quality"
readme = "README.md"
authors = []
requires-python = "==3.12.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "pydantic==2.10.1",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "qdrant-client>=1.12.0",
    "python-dotenv>=1.0.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-typescript>=0.21.0",
    "tree-sitter-rust>=0.21.0",
    "tree-sitter-go>=0.21.0",
    "tree-sitter-java>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
full = [
    "tree-sitter>=0.25.0",
    "tree-sitter-python>=0.25.0",
    "networkx>=3.0",
]

[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "review_eval"
module-root = ""

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
exclude = ["review_eval/fixtures/**", "tests/**"]
venvPath = "."
venv = ".venv"
